# backend/YanCarz/swagger.json
{
    "openapi":  "3.0.4",
    "info":  {
                 "title":  "YanCarz.API",
                 "version":  "1.0"
             },
    "paths":  {
...
                  "/api/admin/User":  {
                                          "get":  {
                                                      "tags":  [
                                                                   "User"
                                                               ],
                                                      "responses":  {
                                                                        "200":  {
                                                                                    "description":  "OK",
                                                                                    "content":  {
                                                                                                    "text/plain":  {
                                                                                                                       "schema":  {
                                                                                                                                      "type":  "array",
                                                                                                                                      "items":  {
                                                                                                                                                    "$ref":  "#/components/schemas/UserDto"
                                                                                                                                                }
                                                                                                                                  }
                                                                                                                   },
                                                                                                    "application/json":  {
                                                                                                                             "schema":  {
                                                                                                                                            "type":  "array",
                                                                                                                                            "items":  {
                                                                                                                                                          "$ref":  "#/components/schemas/UserDto"
                                                                                                                                                      }
                                                                                                                                        }
                                                                                                                         },
                                                                                                    "text/json":  {
                                                                                                                      "schema":  {
                                                                                                                                     "type":  "array",
                                                                                                                                     "items":  {
                                                                                                                                                   "$ref":  "#/components/schemas/UserDto"
                                                                                                                                               }
                                                                                                                                 }
                                                                                                                  }
                                                                                                }
                                                                                }
                                                                    }
                                                  },
                                          "post":  {
                                                       "tags":  [
                                                                    "User"
                                                                ],
                                                       "requestBody":  {
                                                                           "content":  {
                                                                                           "application/json":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/CreateUserRequest"
                                                                                                                               }
                                                                                                                },
                                                                                           "text/json":  {
                                                                                                             "schema":  {
                                                                                                                            "$ref":  "#/components/schemas/CreateUserRequest"
                                                                                                                        }
                                                                                                         },
                                                                                           "application/*+json":  {
                                                                                                                      "schema":  {
                                                                                                                                     "$ref":  "#/components/schemas/CreateUserRequest"
                                                                                                                                 }
                                                                                                                  }
                                                                                       }
                                                                       },
                                                       "responses":  {
                                                                         "200":  {
                                                                                     "description":  "OK"
                                                                                 }
                                                                     }
                                                   }
                                      },
                  "/api/admin/User/{id}":  {
                                               "get":  {
                                                           "tags":  [
                                                                        "User"
                                                                    ],
                                                           "parameters":  [
                                                                              {
                                                                                  "name":  "id",
                                                                                  "in":  "path",
                                                                                  "required":  true,
                                                                                  "schema":  {
                                                                                                 "type":  "string",
                                                                                                 "format":  "uuid"
                                                                                             }
                                                                              }
                                                                          ],
                                                           "responses":  {
                                                                             "200":  {
                                                                                         "description":  "OK",
                                                                                         "content":  {
                                                                                                         "text/plain":  {
                                                                                                                            "schema":  {
                                                                                                                                           "$ref":  "#/components/schemas/UserDto"
                                                                                                                                       }
                                                                                                                        },
                                                                                                         "application/json":  {
                                                                                                                                  "schema":  {
                                                                                                                                                 "$ref":  "#/components/schemas/UserDto"
                                                                                                                                             }
                                                                                                                              },
                                                                                                         "text/json":  {
                                                                                                                           "schema":  {
                                                                                                                                          "$ref":  "#/components/schemas/UserDto"
                                                                                                                                      }
                                                                                                                       }
                                                                                                     }
                                                                                     }
                                                                         }
                                                       },
                                               "put":  {
                                                           "tags":  [
                                                                        "User"
                                                                    ],
                                                           "parameters":  [
                                                                              {
                                                                                  "name":  "id",
                                                                                  "in":  "path",
                                                                                  "required":  true,
                                                                                  "schema":  {
                                                                                                 "type":  "string",
                                                                                                 "format":  "uuid"
                                                                                             }
                                                                              }
                                                                          ],
                                                           "requestBody":  {
                                                                               "content":  {
                                                                                               "application/json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/UserDto"
                                                                                                                                   }
                                                                                                                    },
                                                                                               "text/json":  {
                                                                                                                 "schema":  {
                                                                                                                                "$ref":  "#/components/schemas/UserDto"
                                                                                                                            }
                                                                                                             },
                                                                                               "application/*+json":  {
                                                                                                                          "schema":  {
                                                                                                                                         "$ref":  "#/components/schemas/UserDto"
                                                                                                                                     }
                                                                                                                      }
                                                                                           }
                                                                           },
                                                           "responses":  {
                                                                             "200":  {
                                                                                         "description":  "OK"
                                                                                     }
                                                                         }
                                                       },
                                               "delete":  {
                                                              "tags":  [
                                                                           "User"
                                                                       ],
                                                              "parameters":  [
                                                                                 {
                                                                                     "name":  "id",
                                                                                     "in":  "path",
                                                                                     "required":  true,
                                                                                     "schema":  {
                                                                                                    "type":  "string",
                                                                                                    "format":  "uuid"
                                                                                                }
                                                                                 }
                                                                             ],
                                                              "responses":  {
                                                                                "200":  {
                                                                                            "description":  "OK"
                                                                                        }
                                                                            }
                                                          }
                                           },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "CreateUserRequest":  {
                                                                 "type":  "object",
                                                                 "properties":  {
                                                                                    "name":  {
                                                                                                 "type":  "string",
                                                                                                 "nullable":  true
                                                                                             },
                                                                                    "email":  {
                                                                                                  "type":  "string",
                                                                                                  "nullable":  true
                                                                                              },
                                                                                    "password":  {
                                                                                                     "type":  "string",
                                                                                                     "nullable":  true
                                                                                                 }
                                                                                },
                                                                 "additionalProperties":  false
                                                             },
...
                                       "UserDto":  {
                                                       "type":  "object",
                                                       "properties":  {
                                                                          "id":  {
                                                                                     "type":  "string",
                                                                                     "format":  "uuid"
                                                                                 },
                                                                          "name":  {
                                                                                       "type":  "string",
                                                                                       "nullable":  true
                                                                                   },
                                                                          "email":  {
                                                                                        "type":  "string",
                                                                                        "nullable":  true
                                                                                    }
                                                                      },
                                                       "additionalProperties":  false
                                                   },
...
                                   }
                   },
...
}